Move components to their own classes
